requirements.txt
annotated-types
anyio
asgiref
certifi
charset-normalizer
contourpy
cycler
distro
Django
django-cors-headers
django-filter
djangorestframework
exceptiongroup
fonttools
h11
httpcore
httpx
idna
kiwisolver
matplotlib
numpy
openai
packaging
pandas
pillow
pip
pydantic
pydantic_core
pyparsing
python-dateutil
pytz
PyYAML
requests
SciencePlots
seaborn
setuptools
six
sniffio
sqlparse
tenacity
tqdm
typing_extensions
tzdata
urllib3
wheel
